id int -> Long

+++ src/main/java/com/shophelperjavasb/shophelperjavasb/products/repositories/ProductsRepository.java
@@ -3,5 +3,5 @@ package com.shophelperjavasb.shophelperjavasb.products.repositories;
 import com.shophelperjavasb.shophelperjavasb.products.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
 }
